Made some more changes
The offcanvas sidebar now will display on all pages after the user logs in. One strange error is that if you click the account settings link, it won't let you open the sidebar again. Also, if you log in, you have to go to the profile page, then refresh the page in order for the sidebar to work. I'm not sure what is causing these bugs, hopefully I can figure it out later.