Made some changes to the way settings gets data from Firebase
I added a subscription so that I can now get all the account details and store them locally in the file. I'm also looking into how forms work so that I can get data from the user as input and then save it into the interface, then call updateProfileDetails and store it in the database.