Finished form
The form now displays the current account information, and allows the user to edit and save their changes. Changes are saved on the database. There are only two issues. The first is that even though the email changes, only the original email can be used to sign in. Second is that I don't know how to change password.