Changed the user landing page.
When the user signs in, they now land on the maps page. Also, I found a bug that we may need to address. When the user is not logged in, they can enter /map-select into the url and the system essentially crashes, getting a completely blank screen. In the future, I think we should make this page have the user id in the url, that way we don't run into issues when manually trying to route to the page.