Setup for error messaging
The system will now update the user profile information and the login email properly. Also, as of right now, I have the system display an error modal if the time since login was too long, and no data is updated. This happens because the updateEmail function time sensitive. This may be something we can address in the future, asking the user to sign in again and then updating the profile.